IT-15917: Driver updates for FPGA interface changes (consolidated air+fluid ultrasonic counters). FPGA Versions: Control3 = 0x0E05 SafetyP = 0x0504
Jobs: IT15917_EboxTrima606
[git-p4: depot-paths = "//depot/main/embedded/Common/": change = 73532]

+++ os/cca_pci_intf.c
@@ -87,8 +87,8 @@ void ccaPciShow (void)
              ccaPciData[index].deviceNo,
              ccaPciData[index].funcNo,
              ccaPciData[index].ipin,
-             ccaPciData[index].pBAR0,
-             ccaPciData[index].pBAR1);
+             (int)ccaPciData[index].pBAR0,
+             (int)ccaPciData[index].pBAR1);
    }
 }
